Lint contracts

+++ contracts/BaseSale.sol
@@ -8,7 +8,6 @@ import "./libraries/CommonStructures.sol";
 import "./ExtendableTokenLocker.sol";
 
 contract BaseSale {
-
     using SafeERC20 for IERC20;
     using Address for address;
     using Address for address payable;
@@ -17,11 +16,11 @@ contract BaseSale {
     bool saleForceStarted;
     bool refundEnabled;
     bool public finalized;
-    uint public totalRaised;
-    uint DIVISOR = 10000;
+    uint256 public totalRaised;
+    uint256 DIVISOR = 10000;
 
-    CommonStructures.SaleConfig                     public saleConfig;
-    mapping (address => CommonStructures.UserData)  public userData;
+    CommonStructures.SaleConfig public saleConfig;
+    mapping(address => CommonStructures.UserData) public userData;
 
     ISaleFactory internal saleSpawner;
     address[] internal contributors;
@@ -32,13 +31,18 @@ contract BaseSale {
     ExtendableTokenLocker public lpLocker;
 
     modifier onlySaleCreator {
-        require(msg.sender == saleConfig.creator,"Caller is not sale creator");
+        require(msg.sender == saleConfig.creator, "Caller is not sale creator");
         _;
     }
 
     modifier onlySaleCreatororFactoryOwner {
         //TODO Replace address(0) with a way to get the factory owner
-        require(msg.sender == saleConfig.creator || msg.sender == address(saleSpawner) || msg.sender == saleSpawner.owner(),"Caller is not sale creator or factory allowed");
+        require(
+            msg.sender == saleConfig.creator ||
+                msg.sender == address(saleSpawner) ||
+                msg.sender == saleSpawner.owner(),
+            "Caller is not sale creator or factory allowed"
+        );
         _;
     }
 
@@ -50,43 +54,59 @@ contract BaseSale {
         return totalRaised >= saleConfig.hardCap || finalized;
     }
 
-    function initialize(CommonStructures.SaleConfig calldata saleConfigNew) public {
-        require(!initialized,"Already initialized");
+    function initialize(CommonStructures.SaleConfig calldata saleConfigNew)
+        public
+    {
+        require(!initialized, "Already initialized");
         saleConfig = saleConfigNew;
         router = IUniswapV2Router02(saleConfigNew.router);
         token = IERC20(saleConfig.token);
-        if(saleConfigNew.lpUnlockTime > 0)
-            lpLocker = new ExtendableTokenLocker(token,saleConfigNew.creator, saleConfigNew.lpUnlockTime);
+        if (saleConfigNew.lpUnlockTime > 0)
+            lpLocker = new ExtendableTokenLocker(
+                token,
+                saleConfigNew.creator,
+                saleConfigNew.lpUnlockTime
+            );
         saleSpawner = ISaleFactory(msg.sender);
         initialized = true;
-
     }
 
     receive() external payable {
-        if(msg.sender != address(router)) {
+        if (msg.sender != address(router)) {
             buyTokens();
         }
     }
 
-    function calculateTokensClaimable(uint valueIn) public view returns (uint){
+    function calculateTokensClaimable(uint256 valueIn)
+        public
+        view
+        returns (uint256)
+    {
         //Sale price = (Tokens per ETH * 1e18) / tokens decimals
         return valueIn * saleConfig.salePrice;
     }
 
-    function getRemainingContribution() external view returns (uint) {
+    function getRemainingContribution() external view returns (uint256) {
         return saleConfig.hardCap - totalRaised;
     }
 
     function buyTokens() public payable {
-        require(saleStarted() && !refundEnabled,"Not started yet");
+        require(saleStarted() && !refundEnabled, "Not started yet");
         CommonStructures.UserData storage userDataSender = userData[msg.sender];
         //Check if it surpases max buy
-        require(userDataSender.contributedAmount + msg.value <= saleConfig.maxBuy,"Exceeds max buy");
+        require(
+            userDataSender.contributedAmount + msg.value <= saleConfig.maxBuy,
+            "Exceeds max buy"
+        );
         //If this is a new user add to array of contributors
-        if(userDataSender.contributedAmount == 0) contributors.push(msg.sender);
+        if (userDataSender.contributedAmount == 0)
+            contributors.push(msg.sender);
         //Update contributed amount
         userDataSender.contributedAmount += msg.value;
-        require(totalRaised + msg.value <= saleConfig.hardCap,"HardCap will be reached");
+        require(
+            totalRaised + msg.value <= saleConfig.hardCap,
+            "HardCap will be reached"
+        );
         //Update total raised
         totalRaised += msg.value;
         //Update users tokens they can claim
@@ -94,21 +114,25 @@ contract BaseSale {
     }
 
     function getRefund() external {
-        require(refundEnabled || totalRaised < saleConfig.hardCap,"Refunds not enabled or doesnt pass config");
+        require(
+            refundEnabled || totalRaised < saleConfig.hardCap,
+            "Refunds not enabled or doesnt pass config"
+        );
         CommonStructures.UserData storage userDataSender = userData[msg.sender];
-        require(!userDataSender.refundTaken,"Refund already claimed");
-        require(userDataSender.contributedAmount > 0,"No contribution");
+        require(!userDataSender.refundTaken, "Refund already claimed");
+        require(userDataSender.contributedAmount > 0, "No contribution");
         userDataSender.refundTaken = true;
         payable(msg.sender).sendValue(userDataSender.contributedAmount);
     }
 
-    function  claimTokens() external{
-        require(!refundEnabled,"Refunds enabled");
-        require(finalized,"Sale not finalized yet");
+    function claimTokens() external {
+        require(!refundEnabled, "Refunds enabled");
+        require(finalized, "Sale not finalized yet");
         CommonStructures.UserData storage userDataSender = userData[msg.sender];
-        require(!userDataSender.tokensClaimed,"Tokens already claimed");
-        require(!userDataSender.refundTaken,"Refund was claimed");
-        require(userDataSender.tokensClaimable > 0,"No tokens to claim");
+        require(!userDataSender.tokensClaimed, "Tokens already claimed");
+        require(!userDataSender.refundTaken, "Refund was claimed");
+        require(userDataSender.tokensClaimable > 0, "No tokens to claim");
+
         userDataSender.tokensClaimed = true;
         token.safeTransfer(msg.sender, userDataSender.tokensClaimable);
     }
@@ -118,28 +142,39 @@ contract BaseSale {
         refundEnabled = true;
     }
 
-    function forceStartSale() external onlySaleCreatororFactoryOwner{
+    function forceStartSale() external onlySaleCreatororFactoryOwner {
         saleForceStarted = true;
     }
 
-    function getTokensToAdd(uint value) public view returns (uint) {
+    function getTokensToAdd(uint256 value) public view returns (uint256) {
         //Listing price = (Tokens per ETH * 1e18) / tokens decimals
         return value * saleConfig.listingPrice;
     }
 
     function finalize() external onlySaleCreatororFactoryOwner {
         //Send team their eth
-        if(saleConfig.teamShare >0) {
-            payable(saleConfig.creator).sendValue((totalRaised * saleConfig.teamShare) / 10000);
+        if (saleConfig.teamShare > 0) {
+            payable(saleConfig.creator).sendValue(
+                (totalRaised * saleConfig.teamShare) / 10000
+            );
         }
         //Approve router to spend tokens
         token.safeApprove(address(router), type(uint256).max);
-        uint feeToFactory = (totalRaised * saleSpawner.getETHFee()) / DIVISOR;
-        uint ETHtoAdd = totalRaised - feeToFactory;
+        uint256 feeToFactory =
+            (totalRaised * saleSpawner.getETHFee()) / DIVISOR;
+        uint256 ETHtoAdd = totalRaised - feeToFactory;
         //Add liq as given
-        uint tokensToAdd = getTokensToAdd(ETHtoAdd);
-        router.addLiquidityETH{value : ETHtoAdd}(address(token), tokensToAdd, tokensToAdd, ETHtoAdd, address(lpLocker) != address(0) ? address(lpLocker) : saleConfig.creator, block.timestamp);
+        uint256 tokensToAdd = getTokensToAdd(ETHtoAdd);
+        router.addLiquidityETH{value: ETHtoAdd}(
+            address(token),
+            tokensToAdd,
+            tokensToAdd,
+            ETHtoAdd,
+            address(lpLocker) != address(0)
+                ? address(lpLocker)
+                : saleConfig.creator,
+            block.timestamp
+        );
         finalized = true;
     }
-
 }
