+++ contracts/interfaces/ISaleFactory.sol
@@ -1,6 +1,8 @@
 //SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.3;
+
 interface ISaleFactory {
     function owner() external view returns (address);
-    function getETHFee() external view returns (uint);
+
+    function getETHFee() external view returns (uint256);
 }
